Fixing Images saving

+++ server/src/main/java/eCommerce/com/eCommerce/service/impl/ProductServiceImpl.java
@@ -13,7 +13,9 @@ import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.io.File;
 import java.io.IOException;
+import java.nio.file.Files;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
@@ -25,28 +27,30 @@ public class ProductServiceImpl implements ProductService {
     private final SubcategoryService subcategoryService;
     private final ReviewService reviewService;
     private final TypeService typeService;
-
-    private final CartItemService cartItemService;
-
-
+    private final ShoppingCartService shoppingCartService;
+    private final String FOLDER_PATH = "D:\\java projects\\e_commerce\\server_images";
     private final static Logger LOGGER = LoggerFactory.
             getLogger(ProductService.class);
 
-    public ProductServiceImpl(ProductRepository productRepository, SubcategoryService subcategoryService, ReviewService reviewService, TypeService typeService, CartItemService cartItemService) {
+    public ProductServiceImpl(ProductRepository productRepository, SubcategoryService subcategoryService, ReviewService reviewService, TypeService typeService, ShoppingCartService shoppingCartService) {
         this.productRepository = productRepository;
         this.subcategoryService = subcategoryService;
         this.reviewService = reviewService;
         this.typeService = typeService;
-        this.cartItemService = cartItemService;
+        this.shoppingCartService = shoppingCartService;
+
     }
 
     @Override
     public String saveProduct(ProductRequest productRequest, MultipartFile image) throws IOException {
-        //if there is no image, the imageData will be null
-        byte[] imageData = null;
+        //if there is no image, the imagePath will be null
+        String imagePath = null;
         if(image != null){
-            imageData = image.getBytes();
+            String filePath = FOLDER_PATH + "\\" + image.getOriginalFilename();
+            imagePath = filePath;
+            image.transferTo(new File(filePath));
         }
+
 
         if(productRequest.getTypeId() == null && productRequest.getSubcategoryId() == null){
             throw new IllegalStateException("The product must have a type or a subcategory!");
@@ -63,14 +67,14 @@ public class ProductServiceImpl implements ProductService {
 
         if(productRequest.getTypeId() != null && productRequest.getSubcategoryId() == null){
           var type = typeService.getType(productRequest.getTypeId());
-          var product = saveProductWithType(productRequest, imageData, sku, type);
+          var product = saveProductWithType(productRequest, imagePath, sku, type);
           productRepository.save(product);
 
 
         }
         if(productRequest.getTypeId() == null && productRequest.getSubcategoryId() != null){
             var subcategory = subcategoryService.getSubcategory(productRequest.getSubcategoryId());
-            var product = saveProductWithSubcategory(productRequest, imageData, sku, subcategory);
+            var product = saveProductWithSubcategory(productRequest, imagePath, sku, subcategory);
             productRepository.save(product);;
 
         }
@@ -80,7 +84,7 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public ProductDto getProductById(Long id) {
+    public ProductDto getProductById(Long id) throws IOException {
         var product = productRepository.findById(id).orElseThrow(() -> new ProductNotFoundException("Product not found!"));
         List<Review> productReviews = reviewService.findAllReviewsByProductId(id);
         //incrementing the views of the product
@@ -126,19 +130,35 @@ public class ProductServiceImpl implements ProductService {
     public String deleteProduct(Long id) {
         var product = productRepository.findById(id).orElseThrow(() -> new ProductNotFoundException("Product not found!"));
 
-        productRepository.delete(product);
-        LOGGER.info("Product with id: " + id + " deleted successfully!");
-        return "Product deleted successfully!";
+        //check if product is associated with a cart item
+        var cartItem = product.getCartItem();
+        if(cartItem != null){
+            LOGGER.info("Product is associated with a cart item!");
+            throw new ProductNotFoundException("Somebody has item: "+ product.getProductName() +" in his cart! Try to delete it later or after one week!");
+        }
+        if(cartItem == null && product.getQuantity() > 0) {
+            LOGGER.info("Product has quantity greater than 0!");
+            throw new ProductNotFoundException("Product has quantity greater than 0! First you need to set the quantity to 0! And then you can delete it!");
+        }
+         if(cartItem == null && product.getQuantity() == 0){
+            productRepository.delete(product);
+            LOGGER.info("Product with id: " + id + " deleted successfully!");
+            return "Product deleted successfully!";
+        }
+         return "Product is associated with an order or has quantity greater than 0!";
+
+
     }
 
-    private ProductDto getProductRightProduct(Product product, List<Review> productReviews) {
+    // TODO: 2/20/2024 delete it later but first u have to check for register it with type
+    /*private ProductDto getProductRightProduct(Product product, List<Review> productReviews) {
         if(product.getType() != null){
             return ProductDto.builder()
                     .name(product.getProductName())
                     .description(product.getDescription())
                     .color(product.getColor())
                     .material(product.getMaterial())
-                    .imageData(product.getImageData())
+
                     .typeName(product.getType().getName())
                     .subcategoryName(product.getSubcategory().getName())
                     .categoryName(product.getSubcategory().getCategory().getName())
@@ -154,7 +174,7 @@ public class ProductServiceImpl implements ProductService {
                      .description(product.getDescription())
                      .color(product.getColor())
                      .material(product.getMaterial())
-                     .imageData(product.getImageData())
+
                      .subcategoryName(product.getSubcategory().getName())
                      .categoryName(product.getSubcategory().getCategory().getName())
                      .reviews(productReviews)
@@ -164,6 +184,32 @@ public class ProductServiceImpl implements ProductService {
                      .build();
        }
        return new ProductDto();
+    }*/
+    private ProductDto getProductRightProduct(Product product, List<Review> productReviews) throws IOException {
+        String filePath = product.getImagePath();
+
+        ProductDto.ProductDtoBuilder builder = ProductDto.builder()
+                .name(product.getProductName())
+                .description(product.getDescription())
+                .color(product.getColor())
+                .material(product.getMaterial())
+                .subcategoryName(product.getSubcategory().getName())
+                .categoryName(product.getSubcategory().getCategory().getName())
+                .reviews(productReviews)
+                .price(product.getPrice())
+                .quantityStatus(product.getQuantityStatus())
+                .rating(product.getRating());
+
+        if (product.getType() != null) {
+            builder.typeName(product.getType().getName());
+        }
+        if(product.getImagePath() != null){
+            builder.imageData(Files.readAllBytes(new File(filePath).toPath()));
+        }
+
+
+
+        return builder.build();
     }
 
 
@@ -173,7 +219,7 @@ public class ProductServiceImpl implements ProductService {
                 + productRequest.getMaterial().substring(0, 3)+":"+ UUID.randomUUID().toString();
         return sku;
     }
-    public Product saveProductWithType(ProductRequest productRequest, byte[] imageData, String sku, Type type){
+    public Product saveProductWithType(ProductRequest productRequest, String imagePath, String sku, Type type){
 
         String quantityStatus = checkingQuantityStatus(productRequest.getQuantity());
 
@@ -186,13 +232,13 @@ public class ProductServiceImpl implements ProductService {
                 .views(0L)
                 .quantityStatus(quantityStatus)
                 .quantity(productRequest.getQuantity())
-                .imageData(imageData)
+                .imagePath(imagePath)
                 .sku(sku)
                 .createdAt(LocalDateTime.now())
                 .type(type)
                 .build();
     }
-    public Product saveProductWithSubcategory(ProductRequest productRequest, byte[] imageData, String sku, Subcategory subcategory){
+    public Product saveProductWithSubcategory(ProductRequest productRequest, String imagePath, String sku, Subcategory subcategory){
 
         String quantityStatus = checkingQuantityStatus(productRequest.getQuantity());
 
@@ -205,7 +251,7 @@ public class ProductServiceImpl implements ProductService {
                 .views(0L)
                 .quantityStatus(quantityStatus)
                 .quantity(productRequest.getQuantity())
-                .imageData(imageData)
+                .imagePath(imagePath)
                 .sku(sku)
                 .createdAt(LocalDateTime.now())
                 .subcategory(subcategory)
